MR-176: implementation of route structure
