--- src/routes/activities/activities.js
@@ -1,2 +1,0 @@
-import Activities from 'Views/Admin/Activities';
-export default Activities;
