--- src/routes/activities/formActivities.js
@@ -1,2 +1,0 @@
-import FormActivities from 'Views/Admin/Activities/FormActivity';
-export default FormActivities;
