--- src/routes/admins/admins.js
@@ -1,2 +1,0 @@
-import Admins from 'Views/SuperAdmin/AdminManagement';
-export default Admins;
