--- src/routes/admins/adminsHome.js
@@ -1,2 +1,0 @@
-import AdminsHome from 'Views/Admin';
-export default AdminsHome;
