--- src/routes/admins/formAdmins.js
@@ -1,2 +1,0 @@
-import AdminForm from 'Views/SuperAdmin/AdminManagement/Form';
-export default AdminForm;
