--- src/routes/classes/classes.js
@@ -1,2 +1,0 @@
-import Classes from 'Views/Admin/Classes';
-export default Classes;
