--- src/routes/classes/formClasses.js
@@ -1,2 +1,0 @@
-import FormClasses from 'Views/Admin/Classes/FormClasses';
-export default FormClasses;
