--- src/routes/members/formMembers.js
@@ -1,2 +1,0 @@
-import FormMembers from 'Views/Admin/Members/FormMembers';
-export default FormMembers;
