--- src/routes/members/members.js
@@ -1,2 +1,0 @@
-import Members from 'Views/Admin/Members';
-export default Members;
