--- src/routes/membersUsers/activitiesMembers.js
@@ -1,2 +1,0 @@
-import ActivitiesMember from 'Views/Member/activities/index';
-export default ActivitiesMember;
