--- src/routes/membersUsers/loginMembers.js
@@ -1,2 +1,0 @@
-import LoginMember from 'Views/Member/login';
-export default LoginMember;
