--- src/routes/membersUsers/memberSchedule.js
@@ -1,2 +1,0 @@
-import MemberSchedule from 'Views/Member/schedule';
-export default MemberSchedule;
