--- src/routes/membersUsers/members.js
@@ -1,2 +1,0 @@
-import MemberUser from 'Views/Member';
-export default MemberUser;
