--- src/routes/membersUsers/membershipsMembers.js
@@ -1,2 +1,0 @@
-import MembershipMember from 'Views/Member/memberships';
-export default MembershipMember;
