--- src/routes/membersUsers/signUpMembers.js
@@ -1,2 +1,0 @@
-import SignUpMember from 'Views/Member/signUp';
-export default SignUpMember;
