--- src/routes/subscriptions/formSubscriptions.js
@@ -1,2 +1,0 @@
-import FormSubscriptions from 'Views/Admin/Subscriptions/FormSubscriptions';
-export default FormSubscriptions;
