--- src/routes/subscriptions/subscriptions.js
@@ -1,2 +1,0 @@
-import Subscriptions from 'Views/Admin/Subscriptions';
-export default Subscriptions;
