--- src/routes/superAdmins/formSuperAdmins.js
@@ -1,2 +1,0 @@
-import FormSuperAdmin from 'Views/SuperAdmin/SAManagement/Form';
-export default FormSuperAdmin;
