--- src/routes/superAdmins/superAdmins.js
@@ -1,2 +1,0 @@
-import SuperAdmins from 'Views/SuperAdmin/SAManagement';
-export default SuperAdmins;
