--- src/routes/trainers/formTrainer.js
@@ -1,2 +1,0 @@
-import FormTrainers from 'Views/Admin/Trainers/FormTrainers';
-export default FormTrainers;
