--- src/routes/trainers/trainer.js
@@ -1,2 +1,0 @@
-import Trainers from 'Views/Admin/Trainers';
-export default Trainers;
